+++ ErsatzTV.Application/ProgramSchedules/Commands/ReplaceProgramScheduleItems.cs
@@ -20,7 +20,12 @@ namespace ErsatzTV.Application.ProgramSchedules.Commands
         PlaybackOrder PlaybackOrder,
         int? MultipleCount,
         TimeSpan? PlayoutDuration,
-        bool? OfflineTail,
+        TailMode TailMode,
+        ProgramScheduleItemCollectionType TailCollectionType,
+        int? TailCollectionId,
+        int? TailMultiCollectionId,
+        int? TailSmartCollectionId,
+        int? TailMediaItemId,
         string CustomTitle) : IProgramScheduleItemRequest;
 
     public record ReplaceProgramScheduleItems