+++ ErsatzTV.Infrastructure/Migrations/20211013012823_Add_DurationTail.Designer.cs
@@ -3,14 +3,16 @@ using System;
 using ErsatzTV.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 namespace ErsatzTV.Infrastructure.Migrations
 {
     [DbContext(typeof(TvContext))]
-    partial class TvContextModelSnapshot : ModelSnapshot
+    [Migration("20211013012823_Add_DurationTail")]
+    partial class Add_DurationTail
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -1096,12 +1098,6 @@ namespace ErsatzTV.Infrastructure.Migrations
 
                     b.Property<DateTime>("Finish")
                         .HasColumnType("TEXT");
-
-                    b.Property<DateTime?>("GuideFinish")
-                        .HasColumnType("TEXT");
-
-                    b.Property<bool>("IsFiller")
-                        .HasColumnType("INTEGER");
 
                     b.Property<int>("MediaItemId")
                         .HasColumnType("INTEGER");
@@ -2450,9 +2446,6 @@ namespace ErsatzTV.Infrastructure.Migrations
 
                             b1.Property<DateTime?>("DurationFinish")
                                 .HasColumnType("TEXT");
-
-                            b1.Property<bool>("InDurationFiller")
-                                .HasColumnType("INTEGER");
 
                             b1.Property<bool>("InFlood")
                                 .HasColumnType("INTEGER");