+++ ErsatzTV.Infrastructure/Migrations/20211013025333_Add_PlayoutAnchor_InDurationFiller.Designer.cs
@@ -3,14 +3,16 @@ using System;
 using ErsatzTV.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 namespace ErsatzTV.Infrastructure.Migrations
 {
     [DbContext(typeof(TvContext))]
-    partial class TvContextModelSnapshot : ModelSnapshot
+    [Migration("20211013025333_Add_PlayoutAnchor_InDurationFiller")]
+    partial class Add_PlayoutAnchor_InDurationFiller
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -1096,12 +1098,6 @@ namespace ErsatzTV.Infrastructure.Migrations
 
                     b.Property<DateTime>("Finish")
                         .HasColumnType("TEXT");
-
-                    b.Property<DateTime?>("GuideFinish")
-                        .HasColumnType("TEXT");
-
-                    b.Property<bool>("IsFiller")
-                        .HasColumnType("INTEGER");
 
                     b.Property<int>("MediaItemId")
                         .HasColumnType("INTEGER");